requirements.txt
PySide6>=6.6.0
pynput>=1.7.6
qtawesome>=1.2.0
google-generativeai>=0.3.0
